refactor(sa): Phase 4 slice 7c — the SA Polars namespace becomes plain typed functions across three focused modules
engine/sa/namespace.py (2,153 LOC) deleted; split by cohesion with bodies
verbatim: risk_weights.py (apply_risk_weights + the 16-helper CRR/B31 RW
chain + the scalar dicts), rw_adjustments.py (FCSM / life-insurance /
guarantee substitution / Art. 123B currency-mismatch multiplier — recorded
finding moved verbatim, NOT fixed — / due-diligence override),
factors_output.py (calculate_rwa, supporting factors, audit). Dead
prepare_columns method deleted (zero call sites). The CRM link-ranking
cross-stage site now defers-imports apply_risk_weights directly; the
sa<->crm lazy-import cycle-breakers are preserved on both sides.

90 call sites converted (15 src + 76 test accessors + 14 registration
imports across 11 files); Art. 114/123 citation-hygiene pins re-pointed
(+2 deliberate parametrized cases); 20 citation-snapshot keys re-homed;
LOGGER_REQUIRED_EXEMPT pruned; max_engine_module_loc banked 2,153 -> 1,494.

Gates: unit+contracts 6,067 (+2 = the pin-scope expansion);
acceptance+integration 1,482; parity byte-identical; ruff/ty/arch_check
green.
