+++ tests/contracts/data/citation_snapshot.json
@@ -306,71 +306,71 @@
   "rwa_calc.engine.irb.namespace::IRBLazyFrame.calculate_maturity_adjustment": [
     "CRR Art. 162"
   ],
-  "rwa_calc.engine.sa.namespace::SALazyFrame.apply_currency_mismatch_multiplier": [
-    "PS1/26, paragraph 123B",
-    "PS1/26, paragraph 123B.3"
-  ],
-  "rwa_calc.engine.sa.namespace::SALazyFrame.apply_due_diligence_override": [
-    "PS1/26, paragraph 110A"
-  ],
-  "rwa_calc.engine.sa.namespace::SALazyFrame.apply_fcsm_rw_substitution": [
-    "CRR Art. 222"
-  ],
-  "rwa_calc.engine.sa.namespace::SALazyFrame.apply_guarantee_substitution": [
-    "CRR Art. 213"
-  ],
-  "rwa_calc.engine.sa.namespace::SALazyFrame.apply_life_insurance_rw_mapping": [
-    "CRR Art. 232"
-  ],
-  "rwa_calc.engine.sa.namespace::SALazyFrame.apply_risk_weights": [
-    "CRR Art. 112"
-  ],
-  "rwa_calc.engine.sa.namespace::SALazyFrame.apply_supporting_factors": [
+  "rwa_calc.engine.sa.factors_output::apply_supporting_factors": [
     "CRR Art. 501"
   ],
-  "rwa_calc.engine.sa.namespace::SALazyFrame.calculate_rwa": [
+  "rwa_calc.engine.sa.factors_output::calculate_rwa": [
     "CRR Art. 113"
   ],
-  "rwa_calc.engine.sa.namespace::_apply_b31_risk_weight_overrides": [
+  "rwa_calc.engine.sa.risk_weights::_apply_b31_risk_weight_overrides": [
     "CRR Art. 134",
     "CRR Art. 137"
   ],
-  "rwa_calc.engine.sa.namespace::_apply_crr_risk_weight_overrides": [
+  "rwa_calc.engine.sa.risk_weights::_apply_crr_risk_weight_overrides": [
     "CRR Art. 134",
     "CRR Art. 137"
   ],
-  "rwa_calc.engine.sa.namespace::_apply_defaulted_risk_weight": [
+  "rwa_calc.engine.sa.risk_weights::_apply_defaulted_risk_weight": [
     "CRR Art. 127",
     "PS1/26, paragraph 127"
   ],
-  "rwa_calc.engine.sa.namespace::_b31_append_high_risk_branch": [
+  "rwa_calc.engine.sa.risk_weights::_b31_append_high_risk_branch": [
     "PS1/26, paragraph 128"
   ],
-  "rwa_calc.engine.sa.namespace::_b31_append_real_estate_branches": [
+  "rwa_calc.engine.sa.risk_weights::_b31_append_real_estate_branches": [
     "PS1/26, paragraph 124"
   ],
-  "rwa_calc.engine.sa.namespace::_b31_append_retail_branches": [
+  "rwa_calc.engine.sa.risk_weights::_b31_append_retail_branches": [
     "PS1/26, paragraph 123"
   ],
-  "rwa_calc.engine.sa.namespace::_b31_unrated_cb_rw_expr": [
+  "rwa_calc.engine.sa.risk_weights::_b31_unrated_cb_rw_expr": [
     "CRR Art. 129",
     "PS1/26, paragraph 129"
   ],
-  "rwa_calc.engine.sa.namespace::_crr_append_real_estate_branches": [
+  "rwa_calc.engine.sa.risk_weights::_crr_append_real_estate_branches": [
     "CRR Art. 124"
   ],
-  "rwa_calc.engine.sa.namespace::_crr_append_retail_branches": [
+  "rwa_calc.engine.sa.risk_weights::_crr_append_retail_branches": [
     "CRR Art. 123"
   ],
-  "rwa_calc.engine.sa.namespace::_crr_unrated_cb_rw_expr": [
+  "rwa_calc.engine.sa.risk_weights::_crr_unrated_cb_rw_expr": [
     "CRR Art. 129"
   ],
-  "rwa_calc.engine.sa.namespace::_eca_meip_rw_expr": [
+  "rwa_calc.engine.sa.risk_weights::_eca_meip_rw_expr": [
     "CRR Art. 137"
   ],
-  "rwa_calc.engine.sa.namespace::_prepare_risk_weight_lookup": [
+  "rwa_calc.engine.sa.risk_weights::_prepare_risk_weight_lookup": [
     "PS1/26, paragraph 139",
     "PS1/26, paragraph 122"
+  ],
+  "rwa_calc.engine.sa.risk_weights::apply_risk_weights": [
+    "CRR Art. 112"
+  ],
+  "rwa_calc.engine.sa.rw_adjustments::apply_currency_mismatch_multiplier": [
+    "PS1/26, paragraph 123B",
+    "PS1/26, paragraph 123B.3"
+  ],
+  "rwa_calc.engine.sa.rw_adjustments::apply_due_diligence_override": [
+    "PS1/26, paragraph 110A"
+  ],
+  "rwa_calc.engine.sa.rw_adjustments::apply_fcsm_rw_substitution": [
+    "CRR Art. 222"
+  ],
+  "rwa_calc.engine.sa.rw_adjustments::apply_guarantee_substitution": [
+    "CRR Art. 213"
+  ],
+  "rwa_calc.engine.sa.rw_adjustments::apply_life_insurance_rw_mapping": [
+    "CRR Art. 232"
   ],
   "rwa_calc.engine.securitisation.allocator::SecuritisationAllocator.allocate": [
     "CRR Art. 109",
