+++ tests/unit/test_due_diligence.py
@@ -32,6 +32,7 @@ from rwa_calc.contracts.errors import (
 )
 from rwa_calc.domain.enums import ErrorCategory, ErrorSeverity
 from rwa_calc.engine.sa.calculator import SACalculator
+from rwa_calc.engine.sa.rw_adjustments import apply_due_diligence_override
 
 # =============================================================================
 # Fixtures
@@ -102,7 +103,7 @@ class TestDueDiligenceOverrideApplication:
             risk_weights=[0.50],
             override_rws=[0.75],
         )
-        result = exposures.sa.apply_due_diligence_override(b31_config)
+        result = apply_due_diligence_override(exposures, b31_config)
         df = result.collect()
         assert df["risk_weight"][0] == pytest.approx(0.75)
 
@@ -114,7 +115,7 @@ class TestDueDiligenceOverrideApplication:
             risk_weights=[1.50],
             override_rws=[0.75],
         )
-        result = exposures.sa.apply_due_diligence_override(b31_config)
+        result = apply_due_diligence_override(exposures, b31_config)
         df = result.collect()
         assert df["risk_weight"][0] == pytest.approx(1.50)
 
@@ -126,7 +127,7 @@ class TestDueDiligenceOverrideApplication:
             risk_weights=[0.75],
             override_rws=[0.75],
         )
-        result = exposures.sa.apply_due_diligence_override(b31_config)
+        result = apply_due_diligence_override(exposures, b31_config)
         df = result.collect()
         assert df["risk_weight"][0] == pytest.approx(0.75)
 
@@ -138,7 +139,7 @@ class TestDueDiligenceOverrideApplication:
             risk_weights=[0.50],
             override_rws=[None],
         )
-        result = exposures.sa.apply_due_diligence_override(b31_config)
+        result = apply_due_diligence_override(exposures, b31_config)
         df = result.collect()
         assert df["risk_weight"][0] == pytest.approx(0.50)
 
@@ -148,7 +149,7 @@ class TestDueDiligenceOverrideApplication:
             risk_weights=[0.50, 1.50, 0.75, 1.00],
             override_rws=[0.75, 0.50, None, 2.50],
         )
-        result = exposures.sa.apply_due_diligence_override(b31_config)
+        result = apply_due_diligence_override(exposures, b31_config)
         df = result.collect()
         assert df["risk_weight"][0] == pytest.approx(0.75)  # override applied
         assert df["risk_weight"][1] == pytest.approx(1.50)  # override lower, ignored
@@ -160,7 +161,7 @@ class TestDueDiligenceOverrideApplication:
     ) -> None:
         """No-op when due_diligence_override_rw column is not in the data."""
         exposures = _exposures_without_dd(risk_weights=[0.50, 1.00])
-        result = exposures.sa.apply_due_diligence_override(b31_config)
+        result = apply_due_diligence_override(exposures, b31_config)
         df = result.collect()
         assert df["risk_weight"][0] == pytest.approx(0.50)
         assert df["risk_weight"][1] == pytest.approx(1.00)
@@ -171,7 +172,7 @@ class TestDueDiligenceOverrideApplication:
             risk_weights=[0.50],
             override_rws=[0.75],
         )
-        result = exposures.sa.apply_due_diligence_override(crr_config)
+        result = apply_due_diligence_override(exposures, crr_config)
         df = result.collect()
         assert df["risk_weight"][0] == pytest.approx(0.50)
         assert "due_diligence_override_applied" not in df.columns
@@ -193,7 +194,7 @@ class TestDueDiligenceAuditColumn:
             risk_weights=[0.50],
             override_rws=[0.75],
         )
-        result = exposures.sa.apply_due_diligence_override(b31_config)
+        result = apply_due_diligence_override(exposures, b31_config)
         df = result.collect()
         assert df["due_diligence_override_applied"][0] is True
 
@@ -205,7 +206,7 @@ class TestDueDiligenceAuditColumn:
             risk_weights=[1.50],
             override_rws=[0.75],
         )
-        result = exposures.sa.apply_due_diligence_override(b31_config)
+        result = apply_due_diligence_override(exposures, b31_config)
         df = result.collect()
         assert df["due_diligence_override_applied"][0] is False
 
@@ -217,7 +218,7 @@ class TestDueDiligenceAuditColumn:
             risk_weights=[0.50],
             override_rws=[None],
         )
-        result = exposures.sa.apply_due_diligence_override(b31_config)
+        result = apply_due_diligence_override(exposures, b31_config)
         df = result.collect()
         assert df["due_diligence_override_applied"][0] is False
 
@@ -229,7 +230,7 @@ class TestDueDiligenceAuditColumn:
             risk_weights=[0.50],
             override_rws=[0.75],
         )
-        result = exposures.sa.apply_due_diligence_override(crr_config)
+        result = apply_due_diligence_override(exposures, crr_config)
         df = result.collect()
         assert "due_diligence_override_applied" not in df.columns
 
@@ -238,7 +239,7 @@ class TestDueDiligenceAuditColumn:
     ) -> None:
         """Audit column is not added when override column is absent."""
         exposures = _exposures_without_dd(risk_weights=[0.50])
-        result = exposures.sa.apply_due_diligence_override(b31_config)
+        result = apply_due_diligence_override(exposures, b31_config)
         df = result.collect()
         assert "due_diligence_override_applied" not in df.columns
 
@@ -257,7 +258,7 @@ class TestDueDiligenceWarnings:
         """Warning emitted when due_diligence_performed column is absent under B31."""
         errors: list[CalculationError] = []
         exposures = _exposures_without_dd(risk_weights=[0.50])
-        exposures.sa.apply_due_diligence_override(b31_config, errors=errors)
+        apply_due_diligence_override(exposures, b31_config, errors=errors)
         assert len(errors) == 1
         assert errors[0].code == ERROR_DUE_DILIGENCE_NOT_PERFORMED
 
@@ -267,7 +268,7 @@ class TestDueDiligenceWarnings:
         """Warning has WARNING severity (not ERROR)."""
         errors: list[CalculationError] = []
         exposures = _exposures_without_dd(risk_weights=[0.50])
-        exposures.sa.apply_due_diligence_override(b31_config, errors=errors)
+        apply_due_diligence_override(exposures, b31_config, errors=errors)
         assert errors[0].severity == ErrorSeverity.WARNING
 
     def test_warning_category_is_data_quality(
@@ -276,7 +277,7 @@ class TestDueDiligenceWarnings:
         """Warning category is DATA_QUALITY."""
         errors: list[CalculationError] = []
         exposures = _exposures_without_dd(risk_weights=[0.50])
-        exposures.sa.apply_due_diligence_override(b31_config, errors=errors)
+        apply_due_diligence_override(exposures, b31_config, errors=errors)
         assert errors[0].category == ErrorCategory.DATA_QUALITY
 
     def test_warning_has_regulatory_reference(
@@ -285,7 +286,7 @@ class TestDueDiligenceWarnings:
         """Warning includes Art. 110A regulatory reference."""
         errors: list[CalculationError] = []
         exposures = _exposures_without_dd(risk_weights=[0.50])
-        exposures.sa.apply_due_diligence_override(b31_config, errors=errors)
+        apply_due_diligence_override(exposures, b31_config, errors=errors)
         assert "110A" in (errors[0].regulatory_reference or "")
 
     def test_warning_has_field_name(
@@ -294,7 +295,7 @@ class TestDueDiligenceWarnings:
         """Warning includes the expected field name."""
         errors: list[CalculationError] = []
         exposures = _exposures_without_dd(risk_weights=[0.50])
-        exposures.sa.apply_due_diligence_override(b31_config, errors=errors)
+        apply_due_diligence_override(exposures, b31_config, errors=errors)
         assert errors[0].field_name == "due_diligence_performed"
 
     def test_no_warning_under_crr(
@@ -303,7 +304,7 @@ class TestDueDiligenceWarnings:
         """No warning under CRR even if DD column is absent."""
         errors: list[CalculationError] = []
         exposures = _exposures_without_dd(risk_weights=[0.50])
-        exposures.sa.apply_due_diligence_override(crr_config, errors=errors)
+        apply_due_diligence_override(exposures, crr_config, errors=errors)
         assert len(errors) == 0
 
     def test_no_warning_when_dd_performed_present(
@@ -316,7 +317,7 @@ class TestDueDiligenceWarnings:
             override_rws=[None],
             dd_performed=[True],
         )
-        exposures.sa.apply_due_diligence_override(b31_config, errors=errors)
+        apply_due_diligence_override(exposures, b31_config, errors=errors)
         assert len(errors) == 0
 
     def test_no_warning_when_errors_none(
@@ -325,7 +326,7 @@ class TestDueDiligenceWarnings:
         """No crash when errors parameter is None (unified/branch paths)."""
         exposures = _exposures_without_dd(risk_weights=[0.50])
         # Should not raise — warnings are silently skipped
-        result = exposures.sa.apply_due_diligence_override(b31_config)
+        result = apply_due_diligence_override(exposures, b31_config)
         df = result.collect()
         assert df["risk_weight"][0] == pytest.approx(0.50)
 
@@ -335,7 +336,7 @@ class TestDueDiligenceWarnings:
         """Warning is emitted once per calculation, not per row."""
         errors: list[CalculationError] = []
         exposures = _exposures_without_dd(risk_weights=[0.50, 0.75, 1.00])
-        exposures.sa.apply_due_diligence_override(b31_config, errors=errors)
+        apply_due_diligence_override(exposures, b31_config, errors=errors)
         assert len(errors) == 1
 
 
@@ -355,7 +356,7 @@ class TestDueDiligenceEdgeCases:
             risk_weights=[0.20],
             override_rws=[0.0],
         )
-        result = exposures.sa.apply_due_diligence_override(b31_config)
+        result = apply_due_diligence_override(exposures, b31_config)
         df = result.collect()
         assert df["risk_weight"][0] == pytest.approx(0.20)
 
@@ -367,7 +368,7 @@ class TestDueDiligenceEdgeCases:
             risk_weights=[1.00],
             override_rws=[12.5],
         )
-        result = exposures.sa.apply_due_diligence_override(b31_config)
+        result = apply_due_diligence_override(exposures, b31_config)
         df = result.collect()
         assert df["risk_weight"][0] == pytest.approx(12.5)
 
@@ -384,7 +385,7 @@ class TestDueDiligenceEdgeCases:
             override_rws=[0.75],
             dd_performed=[False],
         )
-        result = exposures.sa.apply_due_diligence_override(b31_config)
+        result = apply_due_diligence_override(exposures, b31_config)
         df = result.collect()
         assert df["risk_weight"][0] == pytest.approx(0.75)
 
@@ -401,7 +402,7 @@ class TestDueDiligenceEdgeCases:
                 "ead_final": [100_000.0],
             }
         ).lazy()
-        result = lf.sa.apply_due_diligence_override(b31_config)
+        result = apply_due_diligence_override(lf, b31_config)
         df = result.collect()
         assert df["exposure_class"][0] == "corporate"
         assert df["ead_final"][0] == pytest.approx(100_000.0)
